Fix eigen integer errors

+++ stan/math/prim/fun/log1p_exp.hpp
@@ -81,7 +81,7 @@ template <typename T,
           require_container_st<is_container, std::is_arithmetic, T>...>
 inline auto log1p_exp(const T& x) {
   return apply_vector_unary<T>::apply(x, [&](const auto& v) {
-    const auto& v_array = v.derived().array();
+    const auto& v_array = v.derived().template cast<const double>().array();
     return (v_array > 0.0).select(v_array + (-v_array).exp().log1p(),
                                   v_array.exp().log1p());
   });
